requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0

# Authentication
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
django-redis==5.4.0

# Celery & Background Tasks
celery==5.3.6
redis==5.0.1

# CORS
django-cors-headers==4.3.1

# API Documentation
drf-spectacular==0.27.1

# Environment Variables
python-dotenv==1.0.0

# Validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# Development
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
flake8==7.0.0
black==24.1.1
mypy==1.8.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring (optional for bonus)
prometheus-client==0.19.0
